requirements.txt
streamlit
authlib
requests
gspread
google-auth
qrcode[pil]
Pillow
